Refactor listing routes and update tests
