+++ backend/routes/contractorRoutes.js
@@ -20,7 +20,7 @@ router.use(authorizeContractor);
 
 // routes
 
-// Allows CO to view all listings in database (Filter feature included)
+// Allows CO to view all homeowner project listings in database (Filter feature included)
 router.post('/listings', viewMultipleListings);
 // Allows updating the contractor's information 
 router.patch('/profile', updateContractorProfile);
